@@ -20,10 +20,10 @@
     }
   },
   "require": {
-    "php": ">=5.3.9",
+    "php": "^5.6||^7.0",
     "ext-mbstring": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "^4.8"
+    "phpunit/phpunit": "^5.7.20||^6.0"
   }
 }
